Add files via upload
The analysis of COVID-19 data in India provides essential insights into the impact, response, and trends related to the pandemic. This analysis is based on a comprehensive dataset that tracks key COVID-19 metrics, including confirmed cases, deaths, recoveries, testing, and vaccination progress.
Key Findings.
🔹 Daily Case Trends:
The analysis reveals daily trends in confirmed cases, deaths, recoveries, and testing.
🔹Geospatial Analysis:
Visualizing the geographic spread of COVID-19 cases on a map highlights regions with varying severity levels.
Color-coded maps provide an immediate understanding of high-risk areas.
🔹Comparative Analysis:
A comparative analysis across Indian states or regions reveals disparities in COVID-19 statistics.
🔹Vaccination Progress (Optional):
If vaccination data is available, insights into vaccination progress and its effect on case numbers are provided.
Monitoring the rate of vaccination is a key component of public health efforts.
✏ Power Query and DAX measures:
To get the metrics in the dashboard, I used Power Query to clean and transform the data, and DAX measures to create the calculations.

📈 Insights:

✅Public health measures: Public health measures, such as lockdowns and social distancing, were effective in reducing the transmission of the virus during the early stages of the pandemic. However, these measures also had a significant economic and social impact.

✅Vaccination: Vaccination is the most effective way to protect against COVID-19. Efforts should be made to increase vaccination rates in all parts of the country.
✅Healthcare infrastructure: The pandemic has exposed the weaknesses in India's healthcare system. Investments should be made to strengthen the healthcare system, including increasing the number of hospital beds and healthcare workers.

Based on the Insight, the following recommendations can be made:
✔ Continue to implement public health measures, such as mask-wearing and social distancing.
✔Promote vaccination, especially among high-risk groups.
✔Increase testing and contact tracing to identify and isolate cases quickly.
✔Improve surveillance and data collection to better understand the spread of the virus and the effectiveness of public health measures.
✔Provide adequate support to healthcare workers, including training, equipment, and mental health support.
It is also important to note that the situation is constantly evolving, and it is important to stay up-to-date on the latest guidance from public health officials.
👏 Appreciation:
I would like to thank InternCareer for the opportunity to work on this project. It was a great learning experience, and I am proud of the dashboard that we created.
Conclusion
By using Power Query, DAX measures, and Power BI, it is possible to create informative and insightful data visualizations that can be used to track the COVID-19 pandemic in India and make informed decisions about public health measures.